Crate rusoto_docdb

Amazon DocumentDB API documentation

If you're using the service, you're probably looking for DocdbClient and Docdb.

Structs

AddTagsToResourceMessage
Represents the input to AddTagsToResource.

ApplyPendingMaintenanceActionMessage
Represents the input to ApplyPendingMaintenanceAction.

ApplyPendingMaintenanceActionResult

AvailabilityZone
Information about an Availability Zone.

CloudwatchLogsExportConfiguration
The configuration setting for the log types to be enabled for export to Amazon CloudWatch Logs for a specific DB instance or DB cluster.

The EnableLogTypes and DisableLogTypes arrays determine which logs are exported (or not exported) to CloudWatch Logs. The values within these arrays depend on the DB engine that is being used.

CopyDBClusterParameterGroupMessage
Represents the input to CopyDBClusterParameterGroup.

CopyDBClusterParameterGroupResult

CopyDBClusterSnapshotMessage
Represents the input to CopyDBClusterSnapshot.

CopyDBClusterSnapshotResult

CreateDBClusterMessage
Represents the input to CreateDBCluster.

CreateDBClusterParameterGroupMessage
Represents the input of CreateDBClusterParameterGroup.

CreateDBClusterParameterGroupResult

CreateDBClusterResult

CreateDBClusterSnapshotMessage
Represents the input of CreateDBClusterSnapshot.

CreateDBClusterSnapshotResult

CreateDBInstanceMessage
Represents the input to CreateDBInstance.

CreateDBInstanceResult

CreateDBSubnetGroupMessage
Represents the input to CreateDBSubnetGroup.

CreateDBSubnetGroupResult

DBCluster
Detailed information about a DB cluster.

DBClusterMember
Contains information about an instance that is part of a DB cluster.

DBClusterMessage
Represents the output of DescribeDBClusters.

DBClusterParameterGroup
Detailed information about a DB cluster parameter group.

DBClusterParameterGroupDetails
Represents the output of DBClusterParameterGroup.

DBClusterParameterGroupNameMessage
Contains the name of a DB cluster parameter group.

DBClusterParameterGroupsMessage
Represents the output of DBClusterParameterGroups.

DBClusterRole
Describes an AWS Identity and Access Management (IAM) role that is associated with a DB cluster.

DBClusterSnapshot
Detailed information about a DB cluster snapshot.

DBClusterSnapshotAttribute
Contains the name and values of a manual DB cluster snapshot attribute.

Manual DB cluster snapshot attributes are used to authorize other AWS accounts to restore a manual DB cluster snapshot.

DBClusterSnapshotAttributesResult
Detailed information about the attributes that are associated with a DB cluster snapshot.

DBClusterSnapshotMessage
Represents the output of DescribeDBClusterSnapshots.

DBEngineVersion
Detailed information about a DB engine version.

DBEngineVersionMessage
Represents the output of DescribeDBEngineVersions.

DBInstance
Detailed information about a DB instance.

DBInstanceMessage
Represents the output of DescribeDBInstances.

DBInstanceStatusInfo
Provides a list of status information for a DB instance.

DBSubnetGroup
Detailed information about a DB subnet group.

DBSubnetGroupMessage
Represents the output of DescribeDBSubnetGroups.

DeleteDBClusterMessage
Represents the input to DeleteDBCluster.

DeleteDBClusterParameterGroupMessage
Represents the input to DeleteDBClusterParameterGroup.

DeleteDBClusterResult

DeleteDBClusterSnapshotMessage
Represents the input to DeleteDBClusterSnapshot.

DeleteDBClusterSnapshotResult

DeleteDBInstanceMessage
Represents the input to DeleteDBInstance.

DeleteDBInstanceResult

DeleteDBSubnetGroupMessage
Represents the input to DeleteDBSubnetGroup.

DescribeDBClusterParameterGroupsMessage
Represents the input to DescribeDBClusterParameterGroups.

DescribeDBClusterParametersMessage
Represents the input to DescribeDBClusterParameters.

DescribeDBClusterSnapshotAttributesMessage
Represents the input to DescribeDBClusterSnapshotAttributes.

DescribeDBClusterSnapshotAttributesResult

DescribeDBClusterSnapshotsMessage
Represents the input to DescribeDBClusterSnapshots.

DescribeDBClustersMessage
Represents the input to DescribeDBClusters.

DescribeDBEngineVersionsMessage
Represents the input to DescribeDBEngineVersions.

DescribeDBInstancesMessage
Represents the input to DescribeDBInstances.

DescribeDBSubnetGroupsMessage
Represents the input to DescribeDBSubnetGroups.

DescribeEngineDefaultClusterParametersMessage
Represents the input to DescribeEngineDefaultClusterParameters.

DescribeEngineDefaultClusterParametersResult

DescribeEventCategoriesMessage
Represents the input to DescribeEventCategories.

DescribeEventsMessage
Represents the input to DescribeEvents.

DescribeOrderableDBInstanceOptionsMessage
Represents the input to DescribeOrderableDBInstanceOptions.

DescribePendingMaintenanceActionsMessage
Represents the input to DescribePendingMaintenanceActions.

DocdbClient
A client for the Amazon DocDB API.

Endpoint
Network information for accessing a DB cluster or DB instance. Client programs must specify a valid endpoint to access these Amazon DocumentDB resources.

EngineDefaults
Contains the result of a successful invocation of the DescribeEngineDefaultClusterParameters operation.

Event
Detailed information about an event.

EventCategoriesMap
An event source type, accompanied by one or more event category names.

EventCategoriesMessage
Represents the output of DescribeEventCategories.

EventsMessage
Represents the output of DescribeEvents.

FailoverDBClusterMessage
Represents the input to FailoverDBCluster.

FailoverDBClusterResult

Filter
A named set of filter values, used to return a more specific list of results. You can use a filter to match a set of resources by specific criteria, such as IDs.

Wildcards are not supported in filters.

ListTagsForResourceMessage
Represents the input to ListTagsForResource.

ModifyDBClusterMessage
Represents the input to ModifyDBCluster.

ModifyDBClusterParameterGroupMessage
Represents the input to ModifyDBClusterParameterGroup.

ModifyDBClusterResult

ModifyDBClusterSnapshotAttributeMessage
Represents the input to ModifyDBClusterSnapshotAttribute.

ModifyDBClusterSnapshotAttributeResult

ModifyDBInstanceMessage
Represents the input to ModifyDBInstance.

ModifyDBInstanceResult

ModifyDBSubnetGroupMessage
Represents the input to ModifyDBSubnetGroup.

ModifyDBSubnetGroupResult

OrderableDBInstanceOption
The options that are available for a DB instance.

OrderableDBInstanceOptionsMessage
Represents the output of DescribeOrderableDBInstanceOptions.

Parameter
Detailed information about an individual parameter.

PendingCloudwatchLogsExports
A list of the log types whose configuration is still pending. These log types are in the process of being activated or deactivated.

PendingMaintenanceAction
Provides information about a pending maintenance action for a resource.

PendingMaintenanceActionsMessage
Represents the output of DescribePendingMaintenanceActions.

PendingModifiedValues
One or more modified settings for a DB instance. These modified settings have been requested, but haven't been applied yet.

RebootDBInstanceMessage
Represents the input to RebootDBInstance.

RebootDBInstanceResult

RemoveTagsFromResourceMessage
Represents the input to RemoveTagsFromResource.

ResetDBClusterParameterGroupMessage
Represents the input to ResetDBClusterParameterGroup.

ResourcePendingMaintenanceActions
Represents the output of ApplyPendingMaintenanceAction.

RestoreDBClusterFromSnapshotMessage
Represents the input to RestoreDBClusterFromSnapshot.

RestoreDBClusterFromSnapshotResult

RestoreDBClusterToPointInTimeMessage
Represents the input to RestoreDBClusterToPointInTime.

RestoreDBClusterToPointInTimeResult

Subnet
Detailed information about a subnet.

Tag
Metadata assigned to an Amazon DocumentDB resource consisting of a key-value pair.

TagListMessage
Represents the output of ListTagsForResource.

UpgradeTarget
The version of the database engine that a DB instance can be upgraded to.

VpcSecurityGroupMembership
Used as a response element for queries on virtual private cloud (VPC) security group membership.

Enums

AddTagsToResourceError
Errors returned by AddTagsToResource

ApplyPendingMaintenanceActionError
Errors returned by ApplyPendingMaintenanceAction

CopyDBClusterParameterGroupError
Errors returned by CopyDBClusterParameterGroup

CopyDBClusterSnapshotError
Errors returned by CopyDBClusterSnapshot

CreateDBClusterError
Errors returned by CreateDBCluster

CreateDBClusterParameterGroupError
Errors returned by CreateDBClusterParameterGroup

CreateDBClusterSnapshotError
Errors returned by CreateDBClusterSnapshot

CreateDBInstanceError
Errors returned by CreateDBInstance

CreateDBSubnetGroupError
Errors returned by CreateDBSubnetGroup

DeleteDBClusterError
Errors returned by DeleteDBCluster

DeleteDBClusterParameterGroupError
Errors returned by DeleteDBClusterParameterGroup

DeleteDBClusterSnapshotError
Errors returned by DeleteDBClusterSnapshot

DeleteDBInstanceError
Errors returned by DeleteDBInstance

DeleteDBSubnetGroupError
Errors returned by DeleteDBSubnetGroup

DescribeDBClusterParameterGroupsError
Errors returned by DescribeDBClusterParameterGroups

DescribeDBClusterParametersError
Errors returned by DescribeDBClusterParameters

DescribeDBClusterSnapshotAttributesError
Errors returned by DescribeDBClusterSnapshotAttributes

DescribeDBClusterSnapshotsError
Errors returned by DescribeDBClusterSnapshots

DescribeDBClustersError
Errors returned by DescribeDBClusters

DescribeDBEngineVersionsError
Errors returned by DescribeDBEngineVersions

DescribeDBInstancesError
Errors returned by DescribeDBInstances

DescribeDBSubnetGroupsError
Errors returned by DescribeDBSubnetGroups

DescribeEngineDefaultClusterParametersError
Errors returned by DescribeEngineDefaultClusterParameters

DescribeEventCategoriesError
Errors returned by DescribeEventCategories

DescribeEventsError
Errors returned by DescribeEvents

DescribeOrderableDBInstanceOptionsError
Errors returned by DescribeOrderableDBInstanceOptions

DescribePendingMaintenanceActionsError
Errors returned by DescribePendingMaintenanceActions

FailoverDBClusterError
Errors returned by FailoverDBCluster

ListTagsForResourceError
Errors returned by ListTagsForResource

ModifyDBClusterError
Errors returned by ModifyDBCluster

ModifyDBClusterParameterGroupError
Errors returned by ModifyDBClusterParameterGroup

ModifyDBClusterSnapshotAttributeError
Errors returned by ModifyDBClusterSnapshotAttribute

ModifyDBInstanceError
Errors returned by ModifyDBInstance

ModifyDBSubnetGroupError
Errors returned by ModifyDBSubnetGroup

RebootDBInstanceError
Errors returned by RebootDBInstance

RemoveTagsFromResourceError
Errors returned by RemoveTagsFromResource

ResetDBClusterParameterGroupError
Errors returned by ResetDBClusterParameterGroup

RestoreDBClusterFromSnapshotError
Errors returned by RestoreDBClusterFromSnapshot

RestoreDBClusterToPointInTimeError
Errors returned by RestoreDBClusterToPointInTime

Traits

Docdb
Trait representing the capabilities of the Amazon DocDB API. Amazon DocDB clients implement this trait.