Struct rusoto_docdb::DBCluster

pub struct DBCluster {
    pub associated_roles: Option<Vec<DBClusterRole>>,
    pub availability_zones: Option<Vec<String>>,
    pub backup_retention_period: Option<i64>,
    pub cluster_create_time: Option<String>,
    pub db_cluster_arn: Option<String>,
    pub db_cluster_identifier: Option<String>,
    pub db_cluster_members: Option<Vec<DBClusterMember>>,
    pub db_cluster_parameter_group: Option<String>,
    pub db_subnet_group: Option<String>,
    pub db_cluster_resource_id: Option<String>,
    pub earliest_restorable_time: Option<String>,
    pub enabled_cloudwatch_logs_exports: Option<Vec<String>>,
    pub endpoint: Option<String>,
    pub engine: Option<String>,
    pub engine_version: Option<String>,
    pub hosted_zone_id: Option<String>,
    pub kms_key_id: Option<String>,
    pub latest_restorable_time: Option<String>,
    pub master_username: Option<String>,
    pub multi_az: Option<bool>,
    pub percent_progress: Option<String>,
    pub port: Option<i64>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub reader_endpoint: Option<String>,
    pub status: Option<String>,
    pub storage_encrypted: Option<bool>,
    pub vpc_security_groups: Option<Vec<VpcSecurityGroupMembership>>,
}

Detailed information about a DB cluster.

Fields
associated_roles: Option<Vec<DBClusterRole>>

Provides a list of the AWS Identity and Access Management (IAM) roles that are associated with the DB cluster. IAM roles that are associated with a DB cluster grant permission for the DB cluster to access other AWS services on your behalf.

availability_zones: Option<Vec<String>>

Provides the list of Amazon EC2 Availability Zones that instances in the DB cluster can be created in.

backup_retention_period: Option<i64>

Specifies the number of days for which automatic DB snapshots are retained.

cluster_create_time: Option<String>

Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

db_cluster_arn: Option<String>

The Amazon Resource Name (ARN) for the DB cluster.

db_cluster_identifier: Option<String>

Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

db_cluster_members: Option<Vec<DBClusterMember>>

Provides the list of instances that make up the DB cluster.

db_cluster_parameter_group: Option<String>

Specifies the name of the DB cluster parameter group for the DB cluster.

db_subnet_group: Option<String>

Specifies information on the subnet group that is associated with the DB cluster, including the name, description, and subnets in the subnet group.

db_cluster_resource_id: Option<String>

The AWS Region-unique, immutable identifier for the DB cluster. This identifier is found in AWS CloudTrail log entries whenever the AWS KMS key for the DB cluster is accessed.

earliest_restorable_time: Option<String>

The earliest time to which a database can be restored with point-in-time restore.

enabled_cloudwatch_logs_exports: Option<Vec<String>>

A list of log types that this DB cluster is configured to export to Amazon CloudWatch Logs.

endpoint: Option<String>

Specifies the connection endpoint for the primary instance of the DB cluster.

engine: Option<String>

Provides the name of the database engine to be used for this DB cluster.

engine_version: Option<String>

Indicates the database engine version.

hosted_zone_id: Option<String>

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

kms_key_id: Option<String>

If StorageEncrypted is true, the AWS KMS key identifier for the encrypted DB cluster.

latest_restorable_time: Option<String>

Specifies the latest time to which a database can be restored with point-in-time restore.

master_username: Option<String>

Contains the master user name for the DB cluster.

multi_az: Option<bool>

Specifies whether the DB cluster has instances in multiple Availability Zones.

percent_progress: Option<String>

Specifies the progress of the operation as a percentage.

port: Option<i64>

Specifies the port that the database engine is listening on.

preferred_backup_window: Option<String>

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

preferred_maintenance_window: Option<String>

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

reader_endpoint: Option<String>

The reader endpoint for the DB cluster. The reader endpoint for a DB cluster load balances connections across the Amazon DocumentDB replicas that are available in a DB cluster. As clients request new connections to the reader endpoint, Amazon DocumentDB distributes the connection requests among the Amazon DocumentDB replicas in the DB cluster. This functionality can help balance your read workload across multiple Amazon DocumentDB replicas in your DB cluster.

If a failover occurs, and the Amazon DocumentDB replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Amazon DocumentDB replicas in the cluster, you can then reconnect to the reader endpoint.

status: Option<String>

Specifies the current state of this DB cluster.

storage_encrypted: Option<bool>

Specifies whether the DB cluster is encrypted.

vpc_security_groups: Option<Vec<VpcSecurityGroupMembership>>

Provides a list of virtual private cloud (VPC) security groups that the DB cluster belongs to.