Struct rusoto_docdb::DocdbClient

pub struct DocdbClient { /* fields omitted */ }

A client for the Amazon DocDB API.

Methods

impl DocdbClient

pub fn new(region: Region) -> DocdbClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> DocdbClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,