Trait rusoto_docdb::Docdb

pub trait Docdb {
    fn add_tags_to_resource(
        &self, 
        input: AddTagsToResourceMessage
    ) -> RusotoFuture<(), AddTagsToResourceError>;

    fn apply_pending_maintenance_action(
        &self, 
        input: ApplyPendingMaintenanceActionMessage
    ) -> RusotoFuture<ApplyPendingMaintenanceActionResult, ApplyPendingMaintenanceActionError>;

    fn copy_db_cluster_parameter_group(
        &self, 
        input: CopyDBClusterParameterGroupMessage
    ) -> RusotoFuture<CopyDBClusterParameterGroupResult, CopyDBClusterParameterGroupError>;

    fn copy_db_cluster_snapshot(
        &self, 
        input: CopyDBClusterSnapshotMessage
    ) -> RusotoFuture<CopyDBClusterSnapshotResult, CopyDBClusterSnapshotError>;

    fn create_db_cluster(
        &self, 
        input: CreateDBClusterMessage
    ) -> RusotoFuture<CreateDBClusterResult, CreateDBClusterError>;

    fn create_db_cluster_parameter_group(
        &self, 
        input: CreateDBClusterParameterGroupMessage
    ) -> RusotoFuture<CreateDBClusterParameterGroupResult, CreateDBClusterParameterGroupError>;

    fn create_db_cluster_snapshot(
        &self, 
        input: CreateDBClusterSnapshotMessage
    ) -> RusotoFuture<CreateDBClusterSnapshotResult, CreateDBClusterSnapshotError>;

    fn create_db_instance(
        &self, 
        input: CreateDBInstanceMessage
    ) -> RusotoFuture<CreateDBInstanceResult, CreateDBInstanceError>;

    fn create_db_subnet_group(
        &self, 
        input: CreateDBSubnetGroupMessage
    ) -> RusotoFuture<CreateDBSubnetGroupResult, CreateDBSubnetGroupError>;

    fn delete_db_cluster(
        &self, 
        input: DeleteDBClusterMessage
    ) -> RusotoFuture<DeleteDBClusterResult, DeleteDBClusterError>;

    fn delete_db_cluster_parameter_group(
        &self, 
        input: DeleteDBClusterParameterGroupMessage
    ) -> RusotoFuture<(), DeleteDBClusterParameterGroupError>;

    fn delete_db_cluster_snapshot(
        &self, 
        input: DeleteDBClusterSnapshotMessage
    ) -> RusotoFuture<DeleteDBClusterSnapshotResult, DeleteDBClusterSnapshotError>;

    fn delete_db_instance(
        &self, 
        input: DeleteDBInstanceMessage
    ) -> RusotoFuture<DeleteDBInstanceResult, DeleteDBInstanceError>;

    fn delete_db_subnet_group(
        &self, 
        input: DeleteDBSubnetGroupMessage
    ) -> RusotoFuture<(), DeleteDBSubnetGroupError>;

    fn describe_db_cluster_parameter_groups(
        &self, 
        input: DescribeDBClusterParameterGroupsMessage
    ) -> RusotoFuture<DBClusterParameterGroupsMessage, DescribeDBClusterParameterGroupsError>;

    fn describe_db_cluster_parameters(
        &self, 
        input: DescribeDBClusterParametersMessage
    ) -> RusotoFuture<DBClusterParameterGroupDetails, DescribeDBClusterParametersError>;

    fn describe_db_cluster_snapshot_attributes(
        &self, 
        input: DescribeDBClusterSnapshotAttributesMessage
    ) -> RusotoFuture<DescribeDBClusterSnapshotAttributesResult, DescribeDBClusterSnapshotAttributesError>;

    fn describe_db_cluster_snapshots(
        &self, 
        input: DescribeDBClusterSnapshotsMessage
    ) -> RusotoFuture<DBClusterSnapshotMessage, DescribeDBClusterSnapshotsError>;

    fn describe_db_clusters(
        &self, 
        input: DescribeDBClustersMessage
    ) -> RusotoFuture<DBClusterMessage, DescribeDBClustersError>;

    fn describe_db_engine_versions(
        &self, 
        input: DescribeDBEngineVersionsMessage
    ) -> RusotoFuture<DBEngineVersionMessage, DescribeDBEngineVersionsError>;

    fn describe_db_instances(
        &self, 
        input: DescribeDBInstancesMessage
    ) -> RusotoFuture<DBInstanceMessage, DescribeDBInstancesError>;

    fn describe_db_subnet_groups(
        &self, 
        input: DescribeDBSubnetGroupsMessage
    ) -> RusotoFuture<DBSubnetGroupMessage, DescribeDBSubnetGroupsError>;

    fn describe_engine_default_cluster_parameters(
        &self, 
        input: DescribeEngineDefaultClusterParametersMessage
    ) -> RusotoFuture<DescribeEngineDefaultClusterParametersResult, DescribeEngineDefaultClusterParametersError>;

    fn describe_event_categories(
        &self, 
        input: DescribeEventCategoriesMessage
    ) -> RusotoFuture<EventCategoriesMessage, DescribeEventCategoriesError>;

    fn describe_events(
        &self, 
        input: DescribeEventsMessage
    ) -> RusotoFuture<EventsMessage, DescribeEventsError>;

    fn describe_orderable_db_instance_options(
        &self, 
        input: DescribeOrderableDBInstanceOptionsMessage
    ) -> RusotoFuture<OrderableDBInstanceOptionsMessage, DescribeOrderableDBInstanceOptionsError>;

    fn describe_pending_maintenance_actions(
        &self, 
        input: DescribePendingMaintenanceActionsMessage
    ) -> RusotoFuture<PendingMaintenanceActionsMessage, DescribePendingMaintenanceActionsError>;

    fn failover_db_cluster(
        &self, 
        input: FailoverDBClusterMessage
    ) -> RusotoFuture<FailoverDBClusterResult, FailoverDBClusterError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceMessage
    ) -> RusotoFuture<TagListMessage, ListTagsForResourceError>;

    fn modify_db_cluster(
        &self, 
        input: ModifyDBClusterMessage
    ) -> RusotoFuture<ModifyDBClusterResult, ModifyDBClusterError>;

    fn modify_db_cluster_parameter_group(
        &self, 
        input: ModifyDBClusterParameterGroupMessage
    ) -> RusotoFuture<DBClusterParameterGroupNameMessage, ModifyDBClusterParameterGroupError>;

    fn modify_db_cluster_snapshot_attribute(
        &self, 
        input: ModifyDBClusterSnapshotAttributeMessage
    ) -> RusotoFuture<ModifyDBClusterSnapshotAttributeResult, ModifyDBClusterSnapshotAttributeError>;

    fn modify_db_instance(
        &self, 
        input: ModifyDBInstanceMessage
    ) -> RusotoFuture<ModifyDBInstanceResult, ModifyDBInstanceError>;

    fn modify_db_subnet_group(
        &self, 
        input: ModifyDBSubnetGroupMessage
    ) -> RusotoFuture<ModifyDBSubnetGroupResult, ModifyDBSubnetGroupError>;

    fn reboot_db_instance(
        &self, 
        input: RebootDBInstanceMessage
    ) -> RusotoFuture<RebootDBInstanceResult, RebootDBInstanceError>;

    fn remove_tags_from_resource(
        &self, 
        input: RemoveTagsFromResourceMessage
    ) -> RusotoFuture<(), RemoveTagsFromResourceError>;

    fn reset_db_cluster_parameter_group(
        &self, 
        input: ResetDBClusterParameterGroupMessage
    ) -> RusotoFuture<DBClusterParameterGroupNameMessage, ResetDBClusterParameterGroupError>;

    fn restore_db_cluster_from_snapshot(
        &self, 
        input: RestoreDBClusterFromSnapshotMessage
    ) -> RusotoFuture<RestoreDBClusterFromSnapshotResult, RestoreDBClusterFromSnapshotError>;

    fn restore_db_cluster_to_point_in_time(
        &self, 
        input: RestoreDBClusterToPointInTimeMessage
    ) -> RusotoFuture<RestoreDBClusterToPointInTimeResult, RestoreDBClusterToPointInTimeError>;
}

Trait representing the capabilities of the Amazon DocDB API. Amazon DocDB clients implement this trait.

Required methods

fn add_tags_to_resource(
&self,
input: AddTagsToResourceMessage
) -> RusotoFuture<(), AddTagsToResourceError>

Adds metadata tags to an Amazon DocumentDB resource. You can use these tags with cost allocation reporting to track costs that are associated with Amazon DocumentDB resources. or in a Condition statement in an AWS Identity and Access Management (IAM) policy for Amazon DocumentDB.

fn apply_pending_maintenance_action(
&self,
input: ApplyPendingMaintenanceActionMessage
) -> RusotoFuture<ApplyPendingMaintenanceActionResult, ApplyPendingMaintenanceActionError>

Applies a pending maintenance action to a resource (for example, to a DB instance).

fn copy_db_cluster_parameter_group(
&self,
input: CopyDBClusterParameterGroupMessage
) -> RusotoFuture<CopyDBClusterParameterGroupResult, CopyDBClusterParameterGroupError>

Copies the specified DB cluster parameter group.

fn copy_db_cluster_snapshot(
&self,
input: CopyDBClusterSnapshotMessage
) -> RusotoFuture<CopyDBClusterSnapshotResult, CopyDBClusterSnapshotError>

Copies a snapshot of a DB cluster.

To copy a DB cluster snapshot from a shared manual DB cluster snapshot, SourceDBClusterSnapshotIdentifier must be the Amazon Resource Name (ARN) of the shared DB cluster snapshot.

To cancel the copy operation after it is in progress, delete the target DB cluster snapshot identified by TargetDBClusterSnapshotIdentifier while that DB cluster snapshot is in the copying status.

fn create_db_cluster(
&self,
input: CreateDBClusterMessage
) -> RusotoFuture<CreateDBClusterResult, CreateDBClusterError>

Creates a new Amazon DocumentDB DB cluster.

fn create_db_cluster_parameter_group(
&self,
input: CreateDBClusterParameterGroupMessage
) -> RusotoFuture<CreateDBClusterParameterGroupResult, CreateDBClusterParameterGroupError>

Creates a new DB cluster parameter group.

Parameters in a DB cluster parameter group apply to all of the instances in a DB cluster.

A DB cluster parameter group is initially created with the default parameters for the database engine used by instances in the DB cluster. To provide custom values for any of the parameters, you must modify the group after you create it. After you create a DB cluster parameter group, you must associate it with your DB cluster. For the new DB cluster parameter group and associated settings to take effect, you must then reboot the DB instances in the DB cluster without failover.

After you create a DB cluster parameter group, you should wait at least 5 minutes before creating your first DB cluster that uses that DB cluster parameter group as the default parameter group. This allows Amazon DocumentDB to fully complete the create action before the DB cluster parameter group is used as the default for a new DB cluster. This step is especially important for parameters that are critical when creating the default database for a DB cluster, such as the character set for the default database defined by the charactersetdatabase parameter.

fn create_db_cluster_snapshot(
&self,
input: CreateDBClusterSnapshotMessage
) -> RusotoFuture<CreateDBClusterSnapshotResult, CreateDBClusterSnapshotError>

Creates a snapshot of a DB cluster.

fn create_db_instance(
&self,
input: CreateDBInstanceMessage
) -> RusotoFuture<CreateDBInstanceResult, CreateDBInstanceError>

Creates a new DB instance.

fn create_db_subnet_group(
&self,
input: CreateDBSubnetGroupMessage
) -> RusotoFuture<CreateDBSubnetGroupResult, CreateDBSubnetGroupError>

Creates a new DB subnet group. DB subnet groups must contain at least one subnet in at least two Availability Zones in the AWS Region.

fn delete_db_cluster(
&self,
input: DeleteDBClusterMessage
) -> RusotoFuture<DeleteDBClusterResult, DeleteDBClusterError>

Deletes a previously provisioned DB cluster. When you delete a DB cluster, all automated backups for that DB cluster are deleted and can't be recovered. Manual DB cluster snapshots of the specified DB cluster are not deleted.

fn delete_db_cluster_parameter_group(
&self,
input: DeleteDBClusterParameterGroupMessage
) -> RusotoFuture<(), DeleteDBClusterParameterGroupError>

Deletes a specified DB cluster parameter group. The DB cluster parameter group to be deleted can't be associated with any DB clusters.

fn delete_db_cluster_snapshot(
&self,
input: DeleteDBClusterSnapshotMessage
) -> RusotoFuture<DeleteDBClusterSnapshotResult, DeleteDBClusterSnapshotError>

Deletes a DB cluster snapshot. If the snapshot is being copied, the copy operation is terminated.

The DB cluster snapshot must be in the available state to be deleted.

fn delete_db_instance(
&self,
input: DeleteDBInstanceMessage
) -> RusotoFuture<DeleteDBInstanceResult, DeleteDBInstanceError>

Deletes a previously provisioned DB instance.

fn delete_db_subnet_group(
&self,
input: DeleteDBSubnetGroupMessage
) -> RusotoFuture<(), DeleteDBSubnetGroupError>

Deletes a DB subnet group.

The specified database subnet group must not be associated with any DB instances.

fn describe_db_cluster_parameter_groups(
&self,
input: DescribeDBClusterParameterGroupsMessage
) -> RusotoFuture<DBClusterParameterGroupsMessage, DescribeDBClusterParameterGroupsError>

Returns a list of DBClusterParameterGroup descriptions. If a DBClusterParameterGroupName parameter is specified, the list contains only the description of the specified DB cluster parameter group.

fn describe_db_cluster_parameters(
&self,
input: DescribeDBClusterParametersMessage
) -> RusotoFuture<DBClusterParameterGroupDetails, DescribeDBClusterParametersError>

Returns the detailed parameter list for a particular DB cluster parameter group.

fn describe_db_cluster_snapshot_attributes(
&self,
input: DescribeDBClusterSnapshotAttributesMessage
) -> RusotoFuture<DescribeDBClusterSnapshotAttributesResult, DescribeDBClusterSnapshotAttributesError>

Returns a list of DB cluster snapshot attribute names and values for a manual DB cluster snapshot.

When you share snapshots with other AWS accounts, DescribeDBClusterSnapshotAttributes returns the restore attribute and a list of IDs for the AWS accounts that are authorized to copy or restore the manual DB cluster snapshot. If all is included in the list of values for the restore attribute, then the manual DB cluster snapshot is public and can be copied or restored by all AWS accounts.

fn describe_db_cluster_snapshots(
&self,
input: DescribeDBClusterSnapshotsMessage
) -> RusotoFuture<DBClusterSnapshotMessage, DescribeDBClusterSnapshotsError>

Returns information about DB cluster snapshots. This API operation supports pagination.

fn describe_db_clusters(
&self,
input: DescribeDBClustersMessage
) -> RusotoFuture<DBClusterMessage, DescribeDBClustersError>

Returns information about provisioned Amazon DocumentDB DB clusters. This API operation supports pagination.

fn describe_db_engine_versions(
&self,
input: DescribeDBEngineVersionsMessage
) -> RusotoFuture<DBEngineVersionMessage, DescribeDBEngineVersionsError>

Returns a list of the available DB engines.

fn describe_db_instances(
&self,
input: DescribeDBInstancesMessage
) -> RusotoFuture<DBInstanceMessage, DescribeDBInstancesError>

Returns information about provisioned Amazon DocumentDB instances. This API supports pagination.

fn describe_db_subnet_groups(
&self,
input: DescribeDBSubnetGroupsMessage
) -> RusotoFuture<DBSubnetGroupMessage, DescribeDBSubnetGroupsError>

Returns a list of DBSubnetGroup descriptions. If a DBSubnetGroupName is specified, the list will contain only the descriptions of the specified DBSubnetGroup.

fn describe_engine_default_cluster_parameters(
&self,
input: DescribeEngineDefaultClusterParametersMessage
) -> RusotoFuture<DescribeEngineDefaultClusterParametersResult, DescribeEngineDefaultClusterParametersError>

Returns the default engine and system parameter information for the cluster database engine.

fn describe_event_categories(
&self,
input: DescribeEventCategoriesMessage
) -> RusotoFuture<EventCategoriesMessage, DescribeEventCategoriesError>

Displays a list of categories for all event source types, or, if specified, for a specified source type.

fn describe_events(
&self,
input: DescribeEventsMessage
) -> RusotoFuture<EventsMessage, DescribeEventsError>

Returns events related to DB instances, DB security groups, DB snapshots, and DB parameter groups for the past 14 days. You can obtain events specific to a particular DB instance, DB security group, DB snapshot, or DB parameter group by providing the name as a parameter. By default, the events of the past hour are returned.

fn describe_orderable_db_instance_options(
&self,
input: DescribeOrderableDBInstanceOptionsMessage
) -> RusotoFuture<OrderableDBInstanceOptionsMessage, DescribeOrderableDBInstanceOptionsError>

Returns a list of orderable DB instance options for the specified engine.

fn describe_pending_maintenance_actions(
&self,
input: DescribePendingMaintenanceActionsMessage
) -> RusotoFuture<PendingMaintenanceActionsMessage, DescribePendingMaintenanceActionsError>

Returns a list of resources (for example, DB instances) that have at least one pending maintenance action.

fn failover_db_cluster(
&self,
input: FailoverDBClusterMessage
) -> RusotoFuture<FailoverDBClusterResult, FailoverDBClusterError>

Forces a failover for a DB cluster.

A failover for a DB cluster promotes one of the Amazon DocumentDB replicas (read-only instances) in the DB cluster to be the primary instance (the cluster writer).

If the primary instance fails, Amazon DocumentDB automatically fails over to an Amazon DocumentDB replica, if one exists. You can force a failover when you want to simulate a failure of a primary instance for testing.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceMessage
) -> RusotoFuture<TagListMessage, ListTagsForResourceError>

Lists all tags on an Amazon DocumentDB resource.

fn modify_db_cluster(
&self,
input: ModifyDBClusterMessage
) -> RusotoFuture<ModifyDBClusterResult, ModifyDBClusterError>

Modifies a setting for an Amazon DocumentDB DB cluster. You can change one or more database configuration parameters by specifying these parameters and the new values in the request.

fn modify_db_cluster_parameter_group(
&self,
input: ModifyDBClusterParameterGroupMessage
) -> RusotoFuture<DBClusterParameterGroupNameMessage, ModifyDBClusterParameterGroupError>

Modifies the parameters of a DB cluster parameter group. To modify more than one parameter, submit a list of the following: ParameterName, ParameterValue, and ApplyMethod. A maximum of 20 parameters can be modified in a single request.

Changes to dynamic parameters are applied immediately. Changes to static parameters require a reboot or maintenance window before the change can take effect.

After you create a DB cluster parameter group, you should wait at least 5 minutes before creating your first DB cluster that uses that DB cluster parameter group as the default parameter group. This allows Amazon DocumentDB to fully complete the create action before the parameter group is used as the default for a new DB cluster. This step is especially important for parameters that are critical when creating the default database for a DB cluster, such as the character set for the default database defined by the charactersetdatabase parameter.

fn modify_db_cluster_snapshot_attribute(
&self,
input: ModifyDBClusterSnapshotAttributeMessage
) -> RusotoFuture<ModifyDBClusterSnapshotAttributeResult, ModifyDBClusterSnapshotAttributeError>

Adds an attribute and values to, or removes an attribute and values from, a manual DB cluster snapshot.

To share a manual DB cluster snapshot with other AWS accounts, specify restore as the AttributeName, and use the ValuesToAdd parameter to add a list of IDs of the AWS accounts that are authorized to restore the manual DB cluster snapshot. Use the value all to make the manual DB cluster snapshot public, which means that it can be copied or restored by all AWS accounts. Do not add the all value for any manual DB cluster snapshots that contain private information that you don't want available to all AWS accounts. If a manual DB cluster snapshot is encrypted, it can be shared, but only by specifying a list of authorized AWS account IDs for the ValuesToAdd parameter. You can't use all as a value for that parameter in this case.

fn modify_db_instance(
&self,
input: ModifyDBInstanceMessage
) -> RusotoFuture<ModifyDBInstanceResult, ModifyDBInstanceError>

Modifies settings for a DB instance. You can change one or more database configuration parameters by specifying these parameters and the new values in the request.

fn modify_db_subnet_group(
&self,
input: ModifyDBSubnetGroupMessage
) -> RusotoFuture<ModifyDBSubnetGroupResult, ModifyDBSubnetGroupError>

Modifies an existing DB subnet group. DB subnet groups must contain at least one subnet in at least two Availability Zones in the AWS Region.

fn reboot_db_instance(
&self,
input: RebootDBInstanceMessage
) -> RusotoFuture<RebootDBInstanceResult, RebootDBInstanceError>

You might need to reboot your DB instance, usually for maintenance reasons. For example, if you make certain changes, or if you change the DB cluster parameter group that is associated with the DB instance, you must reboot the instance for the changes to take effect.

Rebooting a DB instance restarts the database engine service. Rebooting a DB instance results in a momentary outage, during which the DB instance status is set to rebooting.

fn remove_tags_from_resource(
&self,
input: RemoveTagsFromResourceMessage
) -> RusotoFuture<(), RemoveTagsFromResourceError>

Removes metadata tags from an Amazon DocumentDB resource.

fn reset_db_cluster_parameter_group(
&self,
input: ResetDBClusterParameterGroupMessage
) -> RusotoFuture<DBClusterParameterGroupNameMessage, ResetDBClusterParameterGroupError>

Modifies the parameters of a DB cluster parameter group to the default value. To reset specific parameters, submit a list of the following: ParameterName and ApplyMethod. To reset the entire DB cluster parameter group, specify the DBClusterParameterGroupName and ResetAllParameters parameters.

When you reset the entire group, dynamic parameters are updated immediately and static parameters are set to pending-reboot to take effect on the next DB instance reboot.

fn restore_db_cluster_from_snapshot(
&self,
input: RestoreDBClusterFromSnapshotMessage
) -> RusotoFuture<RestoreDBClusterFromSnapshotResult, RestoreDBClusterFromSnapshotError>

Creates a new DB cluster from a DB snapshot or DB cluster snapshot.

If a DB snapshot is specified, the target DB cluster is created from the source DB snapshot with a default configuration and default security group.

If a DB cluster snapshot is specified, the target DB cluster is created from the source DB cluster restore point with the same configuration as the original source DB cluster, except that the new DB cluster is created with the default security group.

fn restore_db_cluster_to_point_in_time(
&self,
input: RestoreDBClusterToPointInTimeMessage
) -> RusotoFuture<RestoreDBClusterToPointInTimeResult, RestoreDBClusterToPointInTimeError>

Restores a DB cluster to an arbitrary point in time. Users can restore to any point in time before LatestRestorableTime for up to BackupRetentionPeriod days. The target DB cluster is created from the source DB cluster with the same configuration as the original DB cluster, except that the new DB cluster is created with the default DB security group.

Loading content...